requirements.txt
# LangChain ecosystem
langchain
langchain-core
langchain-community
langchain-qdrant
langchain-experimental

# Vector DB
qdrant-client
fastembed

# Embeddings
sentence-transformers
huggingface-hub

# LLM providers
openai
groq

# PDF processing
pdfplumber
pymupdf
pypdf
pymupdf4llm
puremagic

# Data / utilities
numpy
pandas
pydantic
python-multipart
python-dotenv

# API & web
fastapi
uvicorn[standard]
httpx

# Database
psycopg2-binary
sqlalchemy
